US5217 - Convert My Serve to module; Module specific routes file; Rename services/directives with server team prefixes
