+++ crossroads.net/app/constants.js
@@ -16,6 +16,7 @@
       GO_VOLUNTEER: 'crossroads.go_volunteer',
       //GROUP_FINDER: 'crossroads.group_finder',
       MEDIA: 'crossroads.media',
+      MY_SERVE: 'crossroads.my_serve',
       LIVE_STREAM: 'crossroads.live_stream',
       GROUP_TOOL: 'crossroads.grouptool',
       MPTOOLS: 'crossroads.mptools',
