+++ crossroads.net/app/my_serve/myserve.controller.js
@@ -8,7 +8,7 @@
     '$rootScope',
     '$window',
     '$log',
-    'filterState',
+    'ServeTeamFilterState',
     'Session',
     'ServeOpportunities',
     'Groups',
