+++ crossroads.net/app/my_serve/myserve.html
@@ -1,7 +1,7 @@
 <div class="serve-signup-wrapper">
   <h1 class="page-header">Sign Up To Serve </h1>
 
-  <p data-ng-show="!serve.showNoOpportunitiesMsg() || !serve.showNoOpportunitiesMsg()">
+  <div class="clearfix" data-ng-show="!serve.showNoOpportunitiesMsg() || !serve.showNoOpportunitiesMsg()">
     <svg viewBox="0 0 32 32" class="icon icon-check-circle">
       <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="#check-circle">
       </use>
@@ -10,33 +10,38 @@
       <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="#minus-circle">
       </use>
     </svg>Signed Up No
-  </p>
 
-  <refine-list last-date="serve.lastDate" serving-days="serve.groups" original="serve.original"> </refine-list>
-
-  <div class="serve-signup-cont clearfix col-sm-9 col-sm-pull-3 soft-ends">
-
-    <div data-ng-show="serve.showNoOpportunitiesMsg() && !serve.filterState.isActive()">
-      <span dynamic-content="$root.MESSAGES.noServingOpportunities.content | html"></span>
-    </div>
-    <div data-ng-show="serve.showNoOpportunitiesMsg() && serve.filterState.isActive()">
-      <span dynamic-content="$root.MESSAGES.servingFilteredToNone.content | html"></span>
-    </div>
-
-    <form name='serveForm'>
-      <div class="serve-day push-bottom" data-ng-repeat="group in serve.groups" ng-show="group.serveTimes.length > 0">
-        <h4>{{::serve.convertToDate(group.day) | date:'EEEE, MMMM dd, yyyy' }}</h4>
-        <serve-tabs opportunity="opp" serve-date="group.day" ng-repeat="opp in group.serveTimes" ng-show="opp.servingTeams.length > 0" >
-        </serve-tabs>
-      </div>
-      <!-- serve day -->
-    </form>
-
-    <div class="row text-center">
-      <div class="col-sm-12">
-        <button class="btn btn-block-mobile btn-standard" ng-show="serve.showButton()"  ng-click="serve.loadNextMonth()" ng-disabled="serve.loadMore"> {{serve.loadText}} </button>
-      </div>
-    </div>
-
+    <button type="button" class="btn btn-primary btn-block-mobile pull-right">Send Message </button>
   </div>
+
+  <div>
+    <serve-team-refine-list last-date="serve.lastDate" serving-days="serve.groups" original="serve.original"> </serve-team-refine-list>
+
+    <div class="serve-signup-cont clearfix col-sm-9 col-sm-pull-3 soft-ends">
+
+      <div data-ng-show="serve.showNoOpportunitiesMsg() && !serve.filterState.isActive()">
+        <span dynamic-content="$root.MESSAGES.noServingOpportunities.content | html"></span>
+      </div>
+      <div data-ng-show="serve.showNoOpportunitiesMsg() && serve.filterState.isActive()">
+        <span dynamic-content="$root.MESSAGES.servingFilteredToNone.content | html"></span>
+      </div>
+
+      <form name='serveForm'>
+        <div class="serve-day soft-bottom push-bottom" data-ng-repeat="group in serve.groups" ng-show="group.serveTimes.length > 0">
+          <h4 class="flush-top">{{::serve.convertToDate(group.day) | date:'EEEE, MMMM dd, yyyy' }}</h4>
+          <serve-tabs opportunity="opp" serve-date="group.day" ng-repeat="opp in group.serveTimes" ng-show="opp.servingTeams.length > 0" >
+          </serve-tabs>
+        </div>
+        <!-- serve day -->
+      </form>
+
+      <div class="row text-center">
+        <div class="col-sm-12">
+          <button class="btn btn-block-mobile btn-standard" ng-show="serve.showButton()"  ng-click="serve.loadNextMonth()" ng-disabled="serve.loadMore"> {{serve.loadText}} </button>
+        </div>
+      </div>
+
+    </div>
+  </div>
+
 </div>
