+++ crossroads.net/app/my_serve/refine/refineList.html
@@ -1,4 +1,4 @@
-<div class="col-sm-3 col-sm-push-9 clearfix serve-filter mobile-hard-left mobile-hard-right">
+<div class="col-sm-3 col-sm-push-9 clearfix serve-filter mobile-hard-left mobile-hard-right soft-ends">
 
   <div ng-click="toggleFilterCollapse()">
