+++ crossroads.net/app/my_serve/serveTeam.directive.js
@@ -9,7 +9,7 @@
     '$log',
     'Session',
     'ServeOpportunities',
-    'Capacity',
+    'OpportunityCapacityService',
     '$modal',
     'growl'
   ];
