+++ crossroads.net/app/routes.js
@@ -456,29 +456,6 @@
           url: '/event-registration-desired',
           templateUrl: 'my_serve/event_registration_mockup_desired.html'
         })
-        .state('serve-signup', {
-          parent: 'noSideBar',
-          url: '/serve-signup',
-          controller: 'MyServeController as serve',
-          templateUrl: 'my_serve/myserve.html',
-          data: {
-            isProtected: true,
-            meta: {
-              title: 'Signup to Serve',
-              description: ''
-            }
-          },
-          resolve: {
-            loggedin: crds_utilities.checkLoggedin,
-            ServeOpportunities: 'ServeOpportunities',
-            $cookies: '$cookies',
-            Groups: function(ServeOpportunities, $cookies) {
-              return ServeOpportunities.ServeDays.query({
-                id: $cookies.get('userId')
-              }).$promise;
-            }
-          }
-        })
         .state('styleguide', {
           parent: 'noHeaderOrFooter',
           url: '/styleguide',
